test(control): added unit tests for websocket control plane basics and restructured proto unit tests - added unit tests for ControlMessage JSON roundtrip: 	- JoinControl 	- ControlMsg 	- ControlAck - added unit tests for channel_key helper - verify control message serialization/deserialization stays stable - verify channel key generation is stable for the same input and different for different channels/servers - moved crate::proto unit tests into a separate file named ut.rs (unit tests)
