+++ src/proto/mod.rs
@@ -1,2 +1,5 @@
 pub mod control_text;
 pub mod packet;
+
+#[cfg(test)]
+mod ut;
